requirements.txt
kafka-python
boto3
python-dateutil